# web/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "acc API",
    "description": "acc 온라인 저지 REST API. 목록은 커서 페이지네이션, 인증은 HttpOnly 세션 쿠키.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/admin/comments/{id}/visibility": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "comment_visibility",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisibilityBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
...
    "/admin/posts/{id}/resolve": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "resolve_typo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/admin/posts/{id}/visibility": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "post_visibility",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisibilityBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/admin/problems": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "admin_list_problems",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "DRAFT, REVIEW or PUBLIC",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminProblemRow"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/problems/{id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Changing the level or visibility recomputes every solver's rating.",
        "operationId": "admin_update_problem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminProblemPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recomputed"
                }
              }
            }
          }
        }
      }
    },
    "/admin/problems/{id}/decision": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Approve (publish with a level) or reject (back to draft with a comment).\nAlso used to publish official problems directly from draft.",
        "operationId": "decide",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Decision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/admin/problems/{id}/rejudge": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Rejudges every submission of a problem. Ratings follow the new verdicts.",
        "operationId": "rejudge_problem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Requeued"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/reports/{id}/resolve": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Closes every open report on the same target.",
        "operationId": "resolve_report",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/admin/submissions/{id}/rejudge": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "rejudge_submission",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Requeued"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "list_users",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Part of a handle or email.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminUserRow"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{id}/suspend": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Suspended users can still sign in but cannot submit or write.",
        "operationId": "suspend",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspendBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/admin/users/{id}/unsuspend": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "unsuspend",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminProblemPatch": {
        "type": "object",
        "properties": {
          "level": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "status": {
            "type": [
              "string",
              "null"
            ],
            "description": "DRAFT, REVIEW or PUBLIC"
          }
        }
      },
      "AdminProblemRow": {
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "source",
          "level",
          "solved_count",
          "testcase_count",
          "created_at"
        ],
        "properties": {
          "author": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "proposed_level": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "solved_count": {
            "type": "integer",
            "format": "int32"
          },
          "source": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "testcase_count": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "validation_status": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "AdminUserRow": {
        "type": "object",
        "required": [
          "id",
          "role",
          "email_verified",
          "suspended",
          "rating",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "email_verified": {
            "type": "boolean"
          },
          "handle": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "role": {
            "type": "string"
          },
          "suspend_reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "suspended": {
            "type": "boolean"
          }
        }
      },
...
      "Decision": {
        "type": "object",
        "required": [
          "approve"
        ],
        "properties": {
          "approve": {
            "type": "boolean"
          },
          "comment": {
            "type": [
              "string",
              "null"
            ]
          },
          "level": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Final level; required when approving."
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. `NOT_FOUND`."
          },
          "location": {
            "type": [
              "string",
              "null"
            ],
            "description": "For `HANDLE_MOVED`: the handle to go to instead."
          },
          "message": {
            "type": "string",
            "description": "Korean message for display."
          }
        }
      },
...
      "Recomputed": {
        "type": "object",
        "required": [
          "users"
        ],
        "properties": {
          "users": {
            "type": "integer",
            "description": "Users whose rating was recomputed.",
            "minimum": 0
          }
        }
      },
...
      "Requeued": {
        "type": "object",
        "required": [
          "submissions"
        ],
        "properties": {
          "submissions": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "ResolveBody": {
        "type": "object",
        "required": [
          "hide"
        ],
        "properties": {
          "hide": {
            "type": "boolean",
            "description": "true hides the target, false restores it."
          }
        }
      },
...
      "SuspendBody": {
        "type": "object",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
...
      "VisibilityBody": {
        "type": "object",
        "required": [
          "hidden"
        ],
        "properties": {
          "hidden": {
            "type": "boolean"
          }
        }
      }
    }
  }
}